+++ src/editor/src/editor/plugin/justifyCenter/index.js
@@ -7,7 +7,7 @@ KISSY.add("editor/plugin/justifyCenter/index", function (S, Editor, justifyCente
         init:function (editor) {
             justifyCenterCmd.init(editor);
             editor.addButton({
-                contentCls:"ke-toolbar-justifyCenter",
+                contentCls:"ks-editor-toolbar-justifyCenter",
                 title:"居中对齐",
                 mode:Editor.WYSIWYG_MODE
             }, {