+++ src/editor/src/editor/plugin/justifyLeft/index.js
@@ -7,7 +7,7 @@ KISSY.add("editor/plugin/justifyLeft/index", function (S, Editor, justifyCenterC
         init:function (editor) {
             justifyCenterCmd.init(editor);
             editor.addButton({
-                contentCls:"ke-toolbar-justifyLeft",
+                contentCls:"ks-editor-toolbar-justifyLeft",
                 title:"左对齐",
                 mode:Editor.WYSIWYG_MODE
             }, {