+++ src/editor/src/editor/plugin/justifyRight/index.js
@@ -7,7 +7,7 @@ KISSY.add("editor/plugin/justifyRight/index", function (S, Editor, justifyCenter
         init:function (editor) {
             justifyCenterCmd.init(editor);
             editor.addButton({
-                contentCls:"ke-toolbar-justifyRight",
+                contentCls:"ks-editor-toolbar-justifyRight",
                 title:"右对齐",
                 mode:Editor.WYSIWYG_MODE
             }, {