+++ src/editor/src/editor/plugin/maximize/cmd.js
@@ -10,11 +10,11 @@ KISSY.add("editor/plugin/maximize/cmd", function (S, Editor) {
         Event = S.Event,
         DOM = S.DOM,
         iframe,
-        MAXIMIZE_TOOLBAR_CLASS = "ke-toolbar-padding",
+        MAXIMIZE_TOOLBAR_CLASS = "ks-editor-toolbar-padding",
         init = function () {
             if (!iframe) {
                 iframe = new Node("<" + "iframe " +
-                    " class='ke-maximize-shim'" +
+                    " class='ks-editor-maximize-shim'" +
                     " style='" +
                     "position:absolute;" +
                     "top:-9999px;" +