+++ src/editor/src/editor/plugin/strikeThrough/index.js
@@ -4,7 +4,7 @@ KISSY.add("editor/plugin/strikeThrough/index", function (S, Editor, ui,cmd) {
             cmd.init(editor);
             editor.addButton({
                 cmdType:"strikeThrough",
-                contentCls:"ke-toolbar-strikeThrough",
+                contentCls:"ks-editor-toolbar-strikeThrough",
                 title:"删除线 "
             }, undefined, ui.Button);
         }