+++ src/editor/src/editor/plugin/undo/index.js
@@ -6,13 +6,13 @@ KISSY.add("editor/plugin/undo/index", function (S, Editor, Btn, cmd) {
                 mode:Editor.WYSIWYG_MODE,
                 title:"撤销",
                 editor:editor,
-                contentCls:"ke-toolbar-undo"
+                contentCls:"ks-editor-toolbar-undo"
             }, undefined, Btn.UndoBtn);
             editor.addButton({
                 mode:Editor.WYSIWYG_MODE,
                 title:"重做",
                 editor:editor,
-                contentCls:"ke-toolbar-redo"
+                contentCls:"ks-editor-toolbar-redo"
             }, undefined, Btn.RedoBtn);
         }
     };