+++ src/editor/src/editor/plugin/video/dialog.js
@@ -10,7 +10,7 @@ KISSY.add("editor/plugin/video/dialog", function (S, Editor, FlashDialog, Select
             "链接： " +
             "" +
             "<input " +
-            "class='ke-video-url ke-input' style='width:410px;" +
+            "class='ks-editor-video-url ks-editor-input' style='width:410px;" +
             MIDDLE + "'/>" +
             "</label>" +
             "</p>" +
@@ -22,7 +22,7 @@ KISSY.add("editor/plugin/video/dialog", function (S, Editor, FlashDialog, Select
             "<input " +
             " data-verify='^(" + DTIP + "|((?!0$)\\d+))?$' " +
             " data-warning='宽度请输入正整数' " +
-            "class='ke-video-width ke-input' " +
+            "class='ks-editor-video-width ks-editor-input' " +
             "style='width:60px;margin-left:2px;" +
             MIDDLE + "' " +
             "/> 像素" +
@@ -34,7 +34,7 @@ KISSY.add("editor/plugin/video/dialog", function (S, Editor, FlashDialog, Select
             " <input " +
             " data-verify='^(" + DTIP + "|((?!0$)\\d+))?$' " +
             " data-warning='高度请输入正整数' " +
-            "class='ke-video-height ke-input' style='width:60px;" +
+            "class='ks-editor-video-height ks-editor-input' style='width:60px;" +
             MIDDLE +
             "'/> 像素" +
             "</label>" +
@@ -42,7 +42,7 @@ KISSY.add("editor/plugin/video/dialog", function (S, Editor, FlashDialog, Select
             "<tr>" +
             "<td>" +
             "<label>对齐： " +
-            "<select class='ke-video-align' title='对齐'>" +
+            "<select class='ks-editor-video-align' title='对齐'>" +
             "<option value='none'>无</option>" +
             "<option value='left'>左对齐</option>" +
             "<option value='right'>右对齐</option>" +
@@ -54,7 +54,7 @@ KISSY.add("editor/plugin/video/dialog", function (S, Editor, FlashDialog, Select
             "" +
             " data-verify='^\\d+$' " +
             " data-warning='间距请输入非负整数' " +
-            "class='ke-video-margin ke-input' style='width:60px;" +
+            "class='ks-editor-video-margin ks-editor-input' style='width:60px;" +
             MIDDLE + "' value='"
             + MARGIN_DEFAULT + "'/> 像素" +
             "</label>" +
@@ -62,9 +62,9 @@ KISSY.add("editor/plugin/video/dialog", function (S, Editor, FlashDialog, Select
             "</table>" +
             "</div>",
         footHtml = "<div style='padding:5px 20px 20px;'><a " +
-            "class='ke-video-ok ke-button' " +
+            "class='ks-editor-video-ok ks-editor-button' " +
             "style='margin-left:40px;margin-right:20px;'>确定</button> " +
-            "<a class='ke-video-cancel ke-button'>取消</a></div>";
+            "<a class='ks-editor-video-cancel ks-editor-button'>取消</a></div>";
 
     function VideoDialog() {
         VideoDialog.superclass.constructor.apply(this, arguments);
@@ -89,13 +89,13 @@ KISSY.add("editor/plugin/video/dialog", function (S, Editor, FlashDialog, Select
             var self = this,
                 d = self.dialog,
                 el = d.get("el");
-            self.dUrl = el.one(".ke-video-url");
-            self.dAlign = Select.decorate(el.one(".ke-video-align"));
-            self.dMargin = el.one(".ke-video-margin");
-            self.dWidth = el.one(".ke-video-width");
-            self.dHeight = el.one(".ke-video-height");
-            var action = el.one(".ke-video-ok"),
-                cancel = el.one(".ke-video-cancel");
+            self.dUrl = el.one(".ks-editor-video-url");
+            self.dAlign = Select.decorate(el.one(".ks-editor-video-align"));
+            self.dMargin = el.one(".ks-editor-video-margin");
+            self.dWidth = el.one(".ks-editor-video-width");
+            self.dHeight = el.one(".ks-editor-video-height");
+            var action = el.one(".ks-editor-video-ok"),
+                cancel = el.one(".ks-editor-video-cancel");
             action.on("click", self._gen, self);
             cancel.on("click", function (ev) {
                 d.hide();