+++ CHANGELOG.md
@@ -66,6 +66,7 @@ History:
  - [*] proxy/scroll.attach changed to proxy/scroll.attachDrag,proxy/scroll.unAttach changed to proxy/scroll.detachDrag
 
  - [!] 删除 uibase 模块，合并 uibase 到 component 模块
+ - [!] 所有组件设置 prefixCls 只作用于组件最外层根节点 class
 
  - [+] 左莫增强 calendar：新的 ui,增加 destroy 方法